tiff support work. including:
1. libtiff incorperation
2. support of tiffs according to libtiff (excluding JPG compressed. currently don't want a jpg library with my lib)
3. colorful bilevel (both bw and grayscale) support added
4. multipage tiffs page selection support edded
5. added test files (from libtiff origina and from filetype site)
6. added two test classes - one for baseline tiff support and another for the specials
7. unrelated - implemented JPEG form xobject retrieval with some basic size information (realized it'd be simpler, since return xobject forms also for tiffs)

+++ PDFWriter/InputFileStream.cpp
@@ -50,5 +50,5 @@ bool InputFileStream::NotEnded()
 void InputFileStream::Skip(LongBufferSizeType inSkipSize)
 {
 	if(mStream)
-		fseek(mStream,inSkipSize,SEEK_CUR);
+		fseek(mStream,(long)inSkipSize,SEEK_CUR);
 }
